Make our custom caching compatible with numba 0.54
Also fail more gracefully with future changes. Fixes #1060
