+++ tests/common/test_numba.py
@@ -9,6 +9,7 @@ import pytest
 import numpy as np
 
 from libertem.common.numba import rmatmul, prime_numba_cache
+from libertem.common.numba.cache import _cached_njit_reg
 
 
 @pytest.mark.with_numba
@@ -72,6 +73,7 @@ def test_numba_cache(tmpdir_factory):
     nbis = glob.glob(pycache_dir + '/*.nbi')
     nbcs = glob.glob(pycache_dir + '/*.nbc')
 
+    assert len(_cached_njit_reg) >= 1
     assert len(nbis) == 1
     assert len(nbcs) == 2
